Inference of model coefficients

This function performs inference on the model coefficient \(\beta\).

cap_beta(Y, X, gamma = NULL, beta = NULL, method = c("asmp", "LLR"), 
    boot = FALSE, sims = 1000, boot.ci.type = c("bca", "perc"), 
    conf.level = 0.95, verbose = TRUE)

Arguments

Y

a data list of length \(n\). Each list element is a \(T\times p\) matrix, the data matrix of \(T\) observations from \(p\) features.

X

a \(n\times q\) data matrix, the covariate matrix of \(n\) subjects with \(q-1\) predictors. The first column is all ones.

gamma

a \(p\)-dimensional vector, the projecting direction \(\gamma\). Default is NULL. If gamma = NULL, an error warning will be returned.

beta

a \(q\)-dimensional vector, the model coefficient \(\beta\). Default is NULL. If beta = NULL, when boot = FALSE, \(\beta\) will be estimated using the provided \(\gamma\).

method

a character of inference method. If method = "asmp", the inference is made based on the asymptotic variance; if method = "LLR", the likelihood ratio test is conducted. When boot = TRUE, this argument is ignored.

boot

a logic variable, whether bootstrap inference is performed.

sims

a numeric value, the number of bootstrap iterations will be performed.

boot.ci.type

a character of the way of calculating bootstrap confidence interval. If boot.ci.type = "bca", the bias corrected confidence interval is returned; if boot.ci.type = "perc", the percentile confidence interval is returned.

conf.level

a numeric value, the designated significance level. Default is \(0.95\).

verbose

a logic variable, whether the bootstrap procedure is printed. Default is TRUE.

Details

Considering \(y_{it}\) are \(p\)-dimensional independent and identically distributed random samples from a multivariate normal distribution with mean zero and covariance matrix \(\Sigma_{i}\). We assume there exits a \(p\)-dimensional vector \(\gamma\) such that \(z_{it}:=\gamma'y_{it}\) satisfies the multiplicative heteroscedasticity: $$\log(\mathrm{Var}(z_{it}))=\log(\gamma'\Sigma_{i}\gamma)=\beta_{0}+x_{i}'\beta_{1},$$ where \(x_{i}\) contains explanatory variables of subject \(i\), and \(\beta_{0}\) and \(\beta_{1}\) are model coefficients.

The \(\beta\) coefficient is estimated by maximizing the likelihood function. The asymptotic variance is obtained based on maximum likelihood estimator theory.

Value

When method = "asmp", the output is a \(q \times 6\) data frame containing the estimate of \(\beta\) coefficient, the asymptotic standard error, the test statistic, the \(p\)-value, and the lower and upper bound of the confidence interval.

When method = "LLR", the output is a \(q \times 3\) data frame containing the estimate of \(\beta\) coefficient, the test statistic, and the \(p\)-value.

When boot = TRUE,

Inference

point estimate of the \(\beta\) coefficient, as well as the corresponding standard error, test statistic, \(p\)-value, and the lower and upper bound of the confidence interval.

beta.boot

the estimate of the \(\beta\) coefficient in each iteration.

References

Zhao et al. (2018) Covariate Assisted Principal Regression for Covariance Matrix Outcomes <doi:10.1101/425033>

Author

Yi Zhao, Johns Hopkins University, <zhaoyi1026@gmail.com>

Bingkai Wang, Johns Hopkins University, <bwang51@jhmi.edu>

Stewart Mostofsky, Johns Hopkins University, <mostofsky@kennedykrieger.org>

Brian Caffo, Johns Hopkins University, <bcaffo@gmail.com>

Xi Luo, Brown University, <xi.rossi.luo@gmail.com>

Examples

#############################################
data(env.example)
X<-get("X",env.example)
Y<-get("Y",env.example)
Phi<-get("Phi",env.example)

# asymptotic variance
re1<-cap_beta(Y,X,gamma=Phi[,2],method=c("asmp"),boot=FALSE)

# likelihood ratio test
re2<-cap_beta(Y,X,gamma=Phi[,2],method=c("LLR"),boot=FALSE)

# bootstrap confidence interval
# \donttest{
re3<-cap_beta(Y,X,gamma=Phi[,2],boot=TRUE,sims=500,verbose=FALSE)
# }
#############################################