Simulated data

"env.example" is an R environment containing the data generated from the proposed model with \(p=2\).

data("env.example")

Format

An R environment

X

a \(n\times q\) data matrix, the covariate matrix of \(n\) subjects with \(q-1\) predictors. The first column is all ones.

Y

a list of length \(n\). Each list element is a \(T\times p\) matrix, the data matrix of \(T\) observations from \(p\) features.

Phi

a \(p\times p\) matrix, the true projection matrix used to generate the data.

beta

a \(q\times p\) matrix, the true coefficient matrix used to generate the data.

Sigma

a \(p\times p\times n\) array, the covariance matrix of the \(n\) subjects.

Details

For subject \(i\) observation \(t\) (\(i=1,\dots,n\), \(t=1,\dots,T\)), \(y_{it}=(y_{it1},\dots,y_{itp})\) was generated from a \(p\)-dimensional normal distribution with mean zero and covariance \(\Sigma\), where $$\Sigma=\Phi\Lambda\Phi,$$ \(\Phi\) is an orthonormal matrix and \(\Lambda=\mathrm{diag}(\lambda_{1},\dots,\lambda_{p})\) is a diagonal matrix. The eigenvalues \(\lambda_{ij}\) (\(j=1,\dots,p\)) satisfies the following log-linear model $$log(\lambda_{ij})=x_{i}^\top\beta_{j},$$ where \(\beta_{j}\) is the \(j\)th column of beta.

Examples

data(env.example)
X<-get("X",env.example)
Y<-get("Y",env.example)